moving code into aare:: namespace
