+++ file_io/include/aare/helpers.hpp
@@ -4,5 +4,8 @@
 #include <filesystem>
 #include <fmt/core.h>
 
+namespace aare {
+
 bool is_master_file(std::filesystem::path fpath);
 
+}
