+++ file_io/src/FileFactory.cpp
@@ -6,6 +6,8 @@
 #include "aare/utils/logger.hpp"
 #include <iostream>
 
+namespace aare {
+
 FileFactory *FileFactory::get_factory(std::filesystem::path fpath) {
     if (fpath.extension() == ".raw" || fpath.extension() == ".json"){
         aare::logger::debug("Loading",fpath.extension(),"file");
@@ -18,11 +20,12 @@ FileFactory *FileFactory::get_factory(std::filesystem::path fpath) {
     // check if extension is numpy
     else if (fpath.extension() == ".npy") {
         aare::logger::debug("Loading numpy file");
-        return new NumpyFileFactory(fpath);
+        return new aare::NumpyFileFactory(fpath);
     }
 
     throw std::runtime_error("Unsupported file type");
 }
 
+} // namespace aare
